it compiles with -03 and -Werror, ship it!

+++ m5/base/str.cc
@@ -122,7 +122,8 @@ __to_number(string value, T &retval)
     static const int bits = sizeof(T) * 8;
     static const T hexmax = maxnum & (((T)1 << (bits - 4 - sign)) - 1);
     static const T octmax = maxnum & (((T)1 << (bits - 3 - sign)) - 1);
-    static const T signmax = (sign) ? maxnum & (((T)1 << (bits - 1)) - 1) : maxnum;
+//     static const T signmax = (sign) ? maxnum & ( ( (T) 1 << (bits - 1) ) - 1) : maxnum; // original code
+    static const T signmax = (sign) ? maxnum & (T) (((unsigned long long) 1 << (bits - 1) ) - 1) : maxnum; // g++ 4.3 fix
     static const T decmax = signmax / 10;
 
 #if 0
